requirements.txt
ddddocr
pillow
msedge-selenium-tools
selenium==3.141
pywin32